requirements.txt
httpx
openai
faiss-cpu
sentence-transformers
hf_xet
aiogram
dotenv